WIP: manage the snackbar's state locally
If a new error is received while the snackbar is open, the snackbar should close
then re-open with the new error message. This behavior is a bit tricky to model
with boolean flags, so I'm trying to model it as a state machine. This commit is
a work in progress as I migrate to the new architecture.

- Remove `currentError` from `<MessageList>`
- Re-write `<ErrorSnackbar>`'s props to accept an array of error messages
- Use react hooks to listen to prop changes
- Use a simple state machine to manage opening and closing the snackbar
- Update or remove outdated tests

Known bug: if you delete an error that is open in the snackbar, instead of just
closing the snackbar (expected behavior), the snackbar re-opens with the next
error in the list.