Fix: snackbar text changes before it closes
The snackbar text was chaning before the close animation finished. This was
particularly apparent when an error was received while the snackbar was open or
when the error that was being displayed was cleared.

I fixed this by using xstate's `context` feature. When an `OPEN` event is
received, the state machine stores the message. When the component is rendered,
it renders the message stored in the state machine's context instead of
rendering the message from the component's prop. This ensures that the new
message won't be displayed until after the snackbar has closed and opened again.

- Update ErrorSnackbar props: This component now takes a message string instead
  of a message object.
- Update tests to match the new props
- Update the error snackbar state machine with the new behavior